==> Scripts/1. Orders, Products Data Import & Descriptive Analysis.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing orders data
#### 3. Creating list of columns to use for analyis
#### 4. Import products data
#### 5. Product Descriptive Statistics

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing orders data

# importing orders dataframe
df = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

### Step 4

# assigning main project path to variable 'path'
path = r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis'

path

### Step 5

# open the os library, find function that joins paths
# this creates one folder path to your specific file

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)
df.head(3)

# 3. Columns List

# creating list of columns to import

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# using vars_list to define which columns to use in dataframe
# column 'eval_set' removed

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)
df.head(3)

### Step 6

# displays the first 5 rows in dataframe
df.head()

# 4. Importing products data

### Step 7

# importing products file

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

### Step 8

# print first 20 rows from products

df_prods.head(20)

# prints last 35 rows from products

df_prods.tail(35)

### Step 9

# print column names from products

df_prods.columns

### Step 10

# rows and columns of products dataframe

df_prods.shape

# 5. Products descriptive statistics

### Step 11a

# descriptive statistics for products

df_prods.describe()

#### Maximum value for "aisle_id" column: 134

### Step 11b

# data types for products dataframe

df_prods.dtypes

#### Data type of "department_id" column is 64-byte integer
