==> Scripts/2. Data Wrangling & Subsetting.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing order data
#### 3. Describing order data
#### 4. Changing order data types and variable names
#### 5. Value counts for hour of day
#### 6. Importing product data and changing data types
#### 7. Importing departments data and transposing dataframe
#### 8. Using data dictionary to find department and product data
#### 9. Retrieving user information and descriptive statistics
#### 10. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing Order data

# importing orders dataframe
df_ords = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

# assigning main project path to variable 'path'
path = r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis'

path

# joining path of main project to order csv file
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# 3. Describing Order data

# returns data types in dataframe
df_ords.dtypes

# list of columns to import
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# importing data set with specified columns to import
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df_ords.head()

# 4. Changing order data types and variable names

# assigning new data types by creating a dictionary
# this will conserve memory when running script

convert_dict = {'order_id': 'str',
                'user_id': 'str',
                'order_number': np.int8,
                'order_dow': np.int8,
                'order_hour_of_day': np.int8,
                'days_since_prior_order': np.float16
               }

# overwriting old dtypes with new ones from dictionary
df_ords = df_ords.astype(convert_dict)
df_ords.dtypes

# changing column name order_dow to orders_day_of_week
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)
df_ords.head()

# Step 5. Value counts for hour of day

### Busiest hour for orders is 10am with 288,418 total orders

# counting the frequency of values in the order_hour_of_day column while retaining NA values

df_ords['order_hour_of_day'].value_counts(dropna = False)

# 6. Importing product data and changing data types

# importing products file
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# assigning new data types by creating a dictionary
# Changing data types for memory conservation

convert_dict_2 = {'product_id': np.int32,
                'aisle_id': np.int32,
                'department_id': np.int32,
                'prices': np.float32
               }

# overwriting old dtypes with new ones from dictionary
df_prods = df_prods.astype(convert_dict_2)
df_prods.dtypes

# 7. Importing departments data and transposing dataframe

# importing departments file
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# overwriting old dataframe to new transposed dataframe
df_dep_t = df_dep.T

# adding index to df_dep_t
df_dep_t.reset_index()

# Take the first row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

# create new df that copies over rows beyond first row, aka 0
df_dep_t_new = df_dep_t[1:]

# set the new_header as the df header
df_dep_t_new.columns = new_header

# 8. Using data dictionary to find department data

### Department_id 4 is produce department

#create data dictionary

data_dict = df_dep_t_new.to_dict('index')
data_dict

# use data dictionary from departments data to identify department_id 4 from product list

print(data_dict.get('4'))

## Subset of breakfast items

# filters for selected element

df_dep_t_new[df_dep_t_new['department']=='breakfast']

# subset created for breakfast items

df_breakfast = df_prods[df_prods['department_id']==14]
df_breakfast

## Dinner party product details (alcohol, deli, beverages, meat/seafood)

# subset of dinner party products

df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

# reset the index so that they're numbered sequentially instead of skipping numbers because
# of the rows that were filtered out
df_dinner_party.reset_index(drop=True).head(30)

# confirms 7650 rows of data in df_dinner_party

df_dinner_party.describe()

# 9. Retrieving user information and descriptive statistics

# user_id 1 order information

df_user1 = df_ords.loc[df_ords['user_id'] == '1']
df_user1

### User 1 data:
#### Number of orders: 11
#### DOW: Average: Monday/Tuesday, Range: Sunday to Wednesday
#### Time: Average: ~10:00am, Earliest: 7:00am, Latest: 4:00pm
#### Days since prior order: Average: 19, Min: 0, Max: 30 

# descriptive statistics for user_id 1

df_user1.describe()

# 10. Exporting data

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
