==> Scripts/3. Data Consistency Checks.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing product and order data
#### 3. Describing order data
#### 4. Detecting mixed data types in order data
#### 5. Detecting and addressing missing values in order data
#### 6. Detecting duplicates in order data
#### 7. Removing missing values and duplicates from products data
#### 8. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing data

# importing products from Original data
df_prods = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

# importing orders dataframe
# index_col = [0] to remove default index column in df
df_ords = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Prepared Data\orders_wrangled.csv', index_col = [0])

# assigning main project path to variable 'path'
path = r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis'

path

# 3. Describing order data

#### Count of all values (except days since prior order): 3.421083e+06
#### Order ID min: 1, max: 3.421083e+06
#### Max order ID matches total counts, but order IDs are not listed in sequential order. How are order IDs assigned to each user?
#### User order number max: 100. Did someone actually order 100 times?

# descriptive statistics for order data

df_ords.describe()

# 4. Detecting mixed data types in order data

# Check for mixed types in order columns
# None detected

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

# 5. Detecting and addressing missing values in order data

# Finding missing values in orders dataframe

df_ords.isnull().sum()

# subset of orders with missing values

df_nan_ords = df_ords[df_ords['days_since_prior_order'].isnull() == True]
df_nan_ords

### The only column with missing values is days_since_prior_order. This may be because a lot of users have only used the app once, so a prior order doesn't exist.

# shape of orders dataframe

df_ords.shape

### Kept null values because this is the initial order for each user. Replaced null with N/A to flag as initial order

# replace nulls with N/A 

df_ords['days_since_prior_order'].fillna("N/A", inplace = True)
df_ords

# 6. Detecting duplicates in order data

# searching for duplicates in orders df
# none found

df_ords[df_ords.duplicated()]

# another way to see a snapshot of data to see if duplicate values are present

df_ords.duplicated().value_counts()

# 7. Removing missing values and duplicates from products data

# removing missing values and creating new dataframe

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# 8. Exporting data

# exporting clean products data frame

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

# exporting clean order data frame

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))
