==> Scripts/4. Combining & Exporting Data.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing original order data, cleaned order data, and cleaned product data
#### 3. Check shape of original order and cleaned order data
#### 4. Merging original and cleaned order data frames
#### 5. Import products checked data
#### 6. Combine orders_products_combined with products data set
#### 7. Exporting df_merged_ord_prod to pkl

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing original order data, cleaned order data, and cleaned product data

# importing order_products_prior from original data
df_ords_prior = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Original Data\order_products_prior.csv', index_col = False)

# importing orders from prepared data
# index_col = [0] to remove default index column in df
df_ords_clean = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Prepared Data\orders_checked.csv', index_col = [0])

# importing products from prepared data
df_prods_clean = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Prepared Data\products_checked.csv', index_col = False)

# assigning main project path to variable 'path'
path = r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis'

# 3. Check shape of original order and cleaned order data

# checking dimensions of imported dataframe
df_ords_prior.shape

# check the output to determine common column names
df_ords_prior.head(1)

# checking shape of cleaned orders data frame
df_ords_clean.shape

# check the output to determine common column names
df_ords_clean.head(1)

# 4. Merging original and cleaned order data frames

#type of join defaults to inner, so it doesn't need to be specified
orders_products_combined = df_ords_clean.merge(df_ords_prior, on = ['order_id'], indicator = True)
orders_products_combined

# used to verify that all data merged
orders_products_combined['_merge'].value_counts()

#removing _merge flag column 
orders_products_combined = df_ords_clean.merge(df_ords_prior, on = ['order_id'], indicator = False)
orders_products_combined

# checking dimensions of imported dataframe
orders_products_combined.shape

# 5. Import products checked data

# importing products from prepared data
df_prods_clean = pd.read_csv(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Prepared Data\products_checked.csv', index_col = 0)

# 6. Combine orders_products_combined with products data set

# checking shape of cleaned products data frame
df_prods_clean.shape

# display .head of products df to see column headers
df_prods_clean.head(1)

# display .head of orders_products_combined df to see column headers
orders_products_combined.head(1)

### Product_id is common column name between both data sets. 
### Inner join performed using product_id as the key so that matching data is present in final data frame.

#inner join products dataset to orders_products_combined using product_id column
# indicator used to verify that merge is successful

df_merged_ord_prod = orders_products_combined.merge(df_prods_clean, on ='product_id', indicator=True)

### Checking counts of merge values to verify that merge was successful

#all values confirmed to be merged

df_merged_ord_prod['_merge'].value_counts()

#removing _merge flag from df

df_merged_ord_prod = orders_products_combined.merge(df_prods_clean, on ='product_id')

# 7. Exporting df_merged_ord_prod to pkl

# Export data to pkl

df_merged_ord_prod.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
