==> Scripts/6. Grouping Data & Aggregating Variables.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing orders_products_merged
#### 3. User_order_number aggregated mean for department_id
#### 4. Create max_order column using transform()
#### 5. Assigning loyalty label to user IDs using loc()
#### 6. Descriptive statistics for prices by loyalty flag
#### 7. Create 'average_price' column for each user_id
#### 8. Create 'median_days_prior_order' column for each user
#### 9. Exporting ords_prods_merge to pkl

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Import orders_products_merged

# importing PKL order_products_combined from prepared data
ords_prods_merge = pd.read_pickle(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged.pkl')

# assigning main project path to variable 'path'
path = r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis'

# 3. User_order_number aggregated mean for department_id

# average user order numbers by dept ID for ENTIRE dataframe
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

# 4. Create max_order column using transform()

# grouping user_id and user_order_number to generate maximum orders for each user
# creating a new column "max_order" to place results of aggregation

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

#adjusting the max rows to be displayed using the head function to assist with quality checks

pd.options.display.max_rows = None

# 5. Assigning loyalty label to user IDs using loc()

# using max_order value to assign loyalty flags to each user
# users with max order greater than 40 are loyal customers

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# users with max orders between 11-40 are regular customers

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

# users with max orders 10 or less are new customers

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

# 6. Descriptive statistics for prices by loyalty flag

ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'count', 'std', 'min', 'max']})

### Loyal customers purchase the lowest priced products on average among all three groups. The standard deviation is the smallest for this group, which indiciates that loyal customers tend to purchase similarly priced items.

### New customers have the least amount of orders (count) but have the highest average for prices. Their standard deviation is the largest indicating more variability in product prices.

### Regular customers have the largest count of orders. Their average price is not as large as new customers, but it isn't far off. Regular customers have a smaller standard deviation than new customers, but it is still much larger than loyal customers. 

### All three groups have users that have purchased the highest priced items and the lowest priced items.

# 7. Create 'average_price' column for each user_id

# calculating the average product price for each user_id

ords_prods_merge['average_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# assigning spending flags to low spenders if average price is less than 10

ords_prods_merge.loc[ords_prods_merge['average_price'] < 10, 'spending_flag'] = 'Low spender'

# assigning spending flags to high spenders if average price is greater than or equal to 10

ords_prods_merge.loc[ords_prods_merge['average_price'] >= 10, 'spending_flag'] = 'High spender'

# checking counts of flags

ords_prods_merge['spending_flag'].value_counts(dropna = False)

# checking dataframe to verify new column and flags are correct
ords_prods_merge.head()

# 8. Create 'median_days_prior_order' column for each user

# calculating the median days since prior order for each user

ords_prods_merge['median_days_prior_order'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# assigning frequency flags to non-frequent customers if days since prior order is greater than 20

ords_prods_merge.loc[ords_prods_merge['median_days_prior_order'] > 20, 'frequency_flag'] = 'Non-frequent customer'

# assigning frequency flags to regular customers if days since prior order is greater than 10 and less than or equal to 20

ords_prods_merge.loc[(ords_prods_merge['median_days_prior_order'] > 10) & (ords_prods_merge['median_days_prior_order'] <= 20), 'frequency_flag'] = 'Regular customer'

# assigning frequency flags to frequent customers if days since prior order is less than or equal to 10

ords_prods_merge.loc[ords_prods_merge['median_days_prior_order'] <= 10, 'frequency_flag'] = 'Frequent customer'

# checking counts of flags

ords_prods_merge['frequency_flag'].value_counts(dropna = False)

### 5 NaN values likely indicate that there are 5 users where the user(s) did not have a prior order

# checking dataframe to verify new column and flags are correct
ords_prods_merge.head()

# 9. Exporting ords_prods_merge to pkl

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
